scripts/requirements.txt
Flask==2.3.0
Flask-CORS==4.0.0
PyJWT==2.8.0
Werkzeug==2.3.0
